Add Vagrant file
This will enable running the application in a 'sandbox' to enable a level playing field.
The reason for this is to allow for easy development and also make sure that the application is ready for any environment.
Ubuntu 14 is chosen as it allows for easy provisioning and installation of required packages.