create authentication forms and models
these will be specifically for auth module. This allows removing this feature if it is not longer required or expanding its functionality
with ease. The structure becomes easier to maintain and understand. This is coupled by the fact that we can now add template and static
resources to this module making it much easier to maintain