Update to docker compose file
Setup of network configuration for the server to connect to database and client to server
the client and server will be on same network for easier communication.

Note, that this configuration does not include a worker as the worker should be started automatically in the background
from the server. Although an easier approach would be having it in a different container, this way the server has access to the
worker more easily and data can flow more seamlessly