add npm-run-all script command
This will allow building the styles under src/styles using Sass pre-processor in production and in development. The running of scripts
remains the same and thus allowing the application to be setup and installed locally remains the same as well.

Running the command will remain
npm run start which will watch-css files for changes and start-js the local server
npm run build remains the same as well, building the final index.css file for production use